int size reduced from 8 to 4 bytes, a simple doubly linked list is added to use it in LRU and LFU cache eviction policies

+++ include/frequency-histogram/frequency_histogram.h
@@ -6,8 +6,8 @@
 
 class FrequencyHistogram {
  public:
-  virtual std::uint64_t Add(const std::string&) = 0;
-  virtual std::uint64_t Estimate(const std::string&) const = 0;
+  virtual std::uint32_t Add(const std::string&) = 0;
+  virtual std::uint32_t Estimate(const std::string&) const = 0;
 };
