+++ include/hash/fnv1a.h
@@ -6,6 +6,6 @@
 #include <cstdint>
 #include <string>
 
-std::uint64_t Fnv1a(const std::string &);
+std::uint32_t Fnv1a(const std::string &);
 
 #endif  // CENG778_PROJECT_INCLUDE_FNV1A_H
